Move TestFullHelpAction to appropriate file
