+++ test/help_test.py
@@ -1,6 +1,8 @@
 import unittest
 
 from parse_this.help.description import _get_default_help_message, prepare_doc
+from test.helpers import Parseable, ParseableWithPrivateMethod
+from test.utils import captured_output
 
 
 def no_docstring():
@@ -151,6 +153,28 @@ class TestHelp(unittest.TestCase):
                 + "I'm multiline too!",
             },
         )
+
+
+class TestFullHelpAction(unittest.TestCase):
+    def test_help_is_complete(self):
+        with captured_output() as (out, _):
+            self.assertRaises(SystemExit, Parseable.parser.parse_args, ["-h"])
+            help_message = out.getvalue()
+        self.assertIn("parseable", help_message)
+        # Private methods and classmethods are not exposed by default
+        self.assertNotIn("private_method", help_message)
+        self.assertNotIn("cls_method", help_message)
+
+    def test_help_is_complete_with_private_method(self):
+        with captured_output() as (out, _):
+            self.assertRaises(
+                SystemExit, ParseableWithPrivateMethod.parser.parse_args, ["-h"]
+            )
+            help_message = out.getvalue()
+        self.assertIn("parseable", help_message)
+        self.assertIn("private_method", help_message)
+        # Classmethods are not exposed by default
+        self.assertNotIn("cls_method", help_message)
 
 
 if __name__ == "__main__":
